fix: merge bug fix

+++ backend/src/main/java/com/bonobono/backend/community/article/service/ArticleCommentService.java
@@ -6,8 +6,8 @@ import com.bonobono.backend.community.article.entity.Article;
 import com.bonobono.backend.community.article.entity.ArticleComment;
 import com.bonobono.backend.community.article.repository.ArticleCommentRepository;
 import com.bonobono.backend.community.article.repository.ArticleRepository;
+import com.bonobono.backend.global.exception.UserNotAuthorizedException;
 import com.bonobono.backend.member.domain.Member;
-import com.bonobono.backend.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -23,33 +23,26 @@ public class ArticleCommentService {
 
     private final ArticleCommentRepository articleCommentRepository;
 
-    private final MemberRepository memberRepository;
-
-    // 댓글 작성하기
+    // 댓글, 대댓글 작성하기
     @Transactional
-    public ArticleCommentResponseDto save(Long articleId, ArticleCommentRequestDto requestDto) {
-        Member member = memberRepository.findById(requestDto.getMemberId())
-                .orElseThrow(()-> new IllegalArgumentException("해당 멤버가 없습니다. id=" + requestDto.getMemberId()));
-
+    public void save(Member member, Long articleId, ArticleCommentRequestDto requestDto) {
         Article article = articleRepository.findById(articleId)
                 .orElseThrow(()-> new IllegalArgumentException("해당 게시글이 없습니다. id=" + articleId));
 
         ArticleComment parentComment = null;
         if (requestDto.getParentCommentId() != null) {
             parentComment = articleCommentRepository.findById(requestDto.getParentCommentId())
-                    .orElseThrow(()-> new IllegalArgumentException("해당 댓글이 없습니다. id=" + articleId));
+                    .orElseThrow(()-> new IllegalArgumentException("해당 댓글이 없습니다. id=" + requestDto.getParentCommentId()));
         }
         ArticleComment articleComment = articleCommentRepository.save(requestDto.toEntity(article, member, parentComment));
         if (parentComment != null){
             parentComment.addChildComment(articleComment);
         }
-        return new ArticleCommentResponseDto(articleComment, member);
-
     }
 
     // 댓글 조회하기
     @Transactional
-    public List<ArticleCommentResponseDto> findByArticleId(Long articleId, Member member){
+    public List<ArticleCommentResponseDto> findByArticleId(Member member, Long articleId){
         return articleCommentRepository.findAllByArticleIdAndParentCommentIsNull(articleId).stream()
                 .map(articleComment -> new ArticleCommentResponseDto(articleComment, member))
                 .collect(Collectors.toList());
@@ -57,27 +50,32 @@ public class ArticleCommentService {
 
     // 댓글 수정하기
     @Transactional
-    public ArticleCommentResponseDto update(Long articleId, Long commentId, ArticleCommentRequestDto requestDto) {
+    public void update(Member member, Long articleId, Long commentId, ArticleCommentRequestDto requestDto) {
         if (!articleRepository.existsById(articleId)) {
             throw new IllegalArgumentException("해당 게시글이 없습니다. id=" + articleId);
         }
         ArticleComment articleComment = articleCommentRepository.findById(commentId)
                 .orElseThrow(()-> new IllegalArgumentException("해당 댓글이 없습니다. id=" + commentId));
-        Member member = memberRepository.findById(requestDto.getMemberId())
-                .orElseThrow(()-> new IllegalArgumentException("해당 멤버가 없습니다. id=" + requestDto.getMemberId()));
-        articleComment.updateComment(requestDto.getContent());
-        return new ArticleCommentResponseDto(articleComment, member);
+        if(member.getId() == articleComment.getMember().getId()) {
+            articleComment.updateComment(requestDto.getContent());
+        } else {
+            throw new UserNotAuthorizedException("해당 유저는 댓글 작성자가 아닙니다.");
+        }
     }
 
     // 댓글 삭제하기
     @Transactional
-    public void delete(Long articleId, Long commentId){
+    public void delete(Member member, Long articleId, Long commentId){
         if (!articleRepository.existsById(articleId)) {
             throw new IllegalArgumentException("해당 게시글이 없습니다. id=" + articleId);
         }
         ArticleComment articleComment = articleCommentRepository.findById(commentId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 댓글이 없습니다. id=" + commentId));
-        articleCommentRepository.delete(articleComment);
+        if(member.getId() == articleComment.getMember().getId()) {
+            articleCommentRepository.delete(articleComment);
+        } else {
+            throw new UserNotAuthorizedException("해당 멤버는 댓글 작성자가 아닙니다.");
+        }
     }
 
 }
